caption-generator/requirements.txt
openai-whisper==20231117
ffmpeg-python==0.2.0
requests==2.31.0
python-dotenv==1.0.0